# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "펫멘토",
        "description": "코멘토 채용 과제",
        "contact": {
            "email": "whdydql123@icloud.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/logout": {
            "post": {
                "tags": [
                    "인증"
                ],
                "summary": "로그아웃",
                "description": "로그아웃 시 요청한 Bearer JWT를 폐기합니다.",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "성공적으로 로그아웃되어 토큰이 폐기되었습니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                },
                                "example": {
                                    "result": "true",
                                    "message": "로그아웃 되었습니다.",
                                    "value": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증 정보가 없는 상태의 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                },
                                "example": {
                                    "result": "false",
                                    "message": "로그인이 필요합니다.",
                                    "value": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer-key": []
                    }
                ]
            }
        },
        "/refresh": {
            "post": {
                "tags": [
                    "인증"
                ],
                "summary": "토큰 갱신",
                "description": "갱신 시 요청한 Bearer JWT를 폐기하고 새로운 토큰을 발급합니다.",
                "operationId": "0c3ea3c61aed881016646810460499e9",
                "responses": {
                    "200": {
                        "description": "성공적으로 갱신되어 새로운 토큰이 발급되었습니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": {
                                        "schema": "TokenRefreshResponse",
                                        "type": "object",
                                        "allOf": [
                                            {
                                                "$ref": "#/components/schemas/BaseResponse"
                                            },
                                            {
                                                "properties": {
                                                    "": {
                                                        "$ref": "#/components/schemas/RefreshToken"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        ]
                                    }
                                },
                                "example": {
                                    "result": "true",
                                    "message": "갱신되었습니다.",
                                    "value": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwMDAvcmVmcmVzaCIsImlhdCI6MTcxMDgyMDQ3NCwiZXhwIjoxNzEwODI0MDg1LCJuYmYiOjE3MTA4MjA0ODUsImp0aSI6IjZXYnZSMFhITTdiTjZYNWMiLCJzdWIiOiI5Yjk3NjIxMy0wZWMzLTQ2NTQtOTIyYS0xNDQxYTJjZTNjOTEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.VLQX1rc47jhptbgTbk8daOoRLbNi3dfap3KMIsZuTkk",
                                        "type": "bearer",
                                        "expires_in": 3600
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증 정보가 없는 상태의 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                },
                                "example": {
                                    "result": "false",
                                    "message": "로그인이 필요합니다.",
                                    "value": []
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer-key": []
                    }
                ]
            }
        },
        "/social/{type}/login": {
            "get": {
                "tags": [
                    "인증"
                ],
                "summary": "소셜 로그인",
                "description": "OAuth에 기반한 써드파티 로그인 창을 오픈합니다. 브라우저를 통해 직접 접속해야 합니다. \n     *          <br />현재는 인증 완료 후 발급된 Access Token과 redirection 파라메터를 함께 요청한 경우 해당 값을 쿼리 파라메터로 담아 코멘토 홈페이지로 이동시킵니다.\n     *          <br />프론트엔드 페이지가 라라벨 내부로 들어오게 되면 Route를 통해 파라메터를 넘기거나, 세션이나 쿠키 등으로 별도로 처리할 수 있을 거 같습니다. \n     *          <br />**접속 URL:** [http://localhost:8000/social/kakao/login](http://localhost:8000/social/kakao/login)",
                "operationId": "f41a026d14072ea27fda96addbd59d8f",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "인증에 활용하고자 하는 소셜 로그인 제공자 종류입니다. 설정을 추가함에 따라 원하는 써드파티를 쉽게 추가할 수 있습니다.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "kakao",
                            "enum": [
                                "kakao"
                            ]
                        }
                    },
                    {
                        "name": "redirection",
                        "in": "query",
                        "description": "인증 완료 후 리디렉션 될 페이지에서 되돌려 주는 값입니다. 리디렉션이 필요한 경우 활용하면 됩니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "요청한 소셜 로그인 제공자로 정상적으로 이동처리 되었습니다."
                    }
                }
            }
        },
        "/social/{type}/callback": {
            "get": {
                "tags": [
                    "인증"
                ],
                "summary": "소셜 로그인 콜백",
                "description": "OAuth에 기반한 써드파티 로그인 후 써드파티 서비스가 쿼리 파라메터와 함께 리디렉션 할 페이지입니다.",
                "operationId": "e3e9dca123cfe3a554425d4054bd52ea",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "인가 코드입니다.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "요청을 식별할 수 있는 상태코드입니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Access Token과 redirection 파라메터를 함께 요청한 경우 해당 값을 쿼리 파라메터로 담아 코멘토 홈페이지로 이동되었습니다."
                    },
                    "400": {
                        "description": "지원하지 않는 소셜 로그인 제공자입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                },
                                "example": {
                                    "result": "false",
                                    "message": "naver 은 지원하지 않는 소셜 로그인 방식입니다.",
                                    "value": []
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "RefreshToken": {
                "title": "RefreshToken",
                "description": "갱신 토큰 모델",
                "properties": {
                    "access_token": {
                        "description": "갱신된 토큰",
                        "type": "string",
                        "format": "string"
                    },
                    "token_type": {
                        "description": "토큰 유형",
                        "type": "string",
                        "format": "string"
                    },
                    "expires_in": {
                        "description": "만료일자",
                        "type": "integer",
                        "format": "int"
                    }
                },
                "type": "object"
            },
            "TokenRefreshResponse": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseResponse"
                    },
                    {
                        "properties": {
                            "": {
                                "$ref": "#/components/schemas/RefreshToken"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "BaseResponse": {
                "title": "BaseResponse",
                "description": "공통 응답 모델",
                "required": [
                    "result",
                    "message"
                ],
                "properties": {
                    "result": {
                        "description": "응답 결과",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "message": {
                        "description": "응답 메세지",
                        "type": "string",
                        "format": "string"
                    },
                    "value": {
                        "description": "응답",
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer-key": {
                "type": "http",
                "name": "bearer-key",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}